requirements.txt
streamlit>=1.36.0
mistralai>=0.4.0
requests>=2.32.0
pandas>=2.2.2
python-dotenv>=1.0.1
